requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
backports.asyncio.runner==1.2.0
black==25.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
Django==5.2.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.0
fastapi==0.116.1
flake8==7.3.0
gunicorn==23.0.0
h11==0.16.0
idna==3.10
iniconfig==2.1.0
isort==6.0.1
json5==0.12.1
markdown-it-py==4.0.0
mbake==1.4.1
mccabe==0.7.0
mdurl==0.1.2
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-asyncio==1.1.0
python-dotenv==1.1.1
requests==2.32.5
rich==14.1.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
starlette==0.47.2
tomli==2.2.1
typer==0.17.3
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
